Issue #000 fix: git clone

+++ pipelines/player-cdn-upload/setupPlayer.sh
@@ -4,7 +4,10 @@ set -ex
 
 commit_ref=${1:-angular-migration}
 player_dir=sunbird-portal
-[ -d $player_dir ] || git clone https://github.com/project-sunbird/sunbird-player -b $commit_ref && exit 0
+if [ ! -d $player_dir ]; 
+    git clone https://github.com/project-sunbird/sunbird-player -b $commit_ref
+    exit 0
+fi
 cd $player_dir
 git reset --hard
 git clean -fd
